requirements.txt
asgiref==3.6.0
boto3==1.26.146
botocore==1.29.146
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.54.0
django-autoslug==1.9.9
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-storages==1.13.2
django-taggit==3.1.0
idna==3.4
jmespath==1.0.1
Markdown==3.4.3
oauthlib==3.2.2
Pillow==9.5.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.16
